areteia_ai/requirements.txt
fastapi
uvicorn
python-multipart
faiss-cpu
sentence-transformers
pypdf
python-docx
pydantic
python-docx
python-pptx
PyPDF2
langchain==0.0.202
